bonus_comp

+++ libft/ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: sesim <sesim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/10 16:46:09 by sesim             #+#    #+#             */
-/*   Updated: 2022/03/21 14:39:27 by sesim            ###   ########.fr       */
+/*   Updated: 2022/03/25 20:58:58 by seongmins        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	nb_size(int n)
+static size_t	nb_size(int n)
 {
 	int	len;
