+++ libft/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: sesim <sesim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/10 16:49:48 by sesim             #+#    #+#             */
-/*   Updated: 2022/03/24 14:09:22 by sesim            ###   ########.fr       */
+/*   Updated: 2022/03/25 21:28:45 by seongmins        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,6 @@ char	*ft_strtrim(char const *s1, char const *set)
 	while (len > i && ft_strchr(set, s1[len]))
 		len--;
 	len = len - i + 1;
-	res = ft_substr(s1, i, len);
+	res = ft_substr(s1, (unsigned int)i, len);
 	return (res);
 }
